requirements.txt
asgiref==3.7.2
Django==5.0.2
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
phonenumberslite==8.13.31
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
